feat(kubelet, containerd, docker): rename component IDs, names
- kubelet -> kubelet
- containerd-pod -> containerd
- docker-container -> docker
