+++ components/containerd/component_test.go
@@ -1,4 +1,4 @@
-package pod
+package containerd
 
 import (
 	"context"
