--- components/containerd/doc.go
@@ -1,2 +1,0 @@
-// Package containerd contains the containerd components and its query interface.
-package containerd
