+++ components/docker/component_test.go
@@ -1,4 +1,4 @@
-package container
+package docker
 
 import (
 	"context"
