--- components/docker/doc.go
@@ -1,2 +1,0 @@
-// Package docker contains the docker components and its query interface.
-package docker
