+++ components/kubelet/kubelet-readonly-pods.json
